requirements.txt
finance-datareader
langchain 
langgraph 
langchain-openai 
gpt-researcher 
open-interpreter